gnu: home-services: ssh: Use ‘define-configuration’ and improve docs.
* gnu/home-services/ssh.scm (serialize-alist): Take optional keyword
argument ‘toplevel?’.
(serialize-toplevel-alist): Remove procedure.
(<ssh-host>): Use ‘define-configuration’
instead of ‘define-record-type’.
(<ssh-match>): Likewise.
(ssh-host-or-ssh-match?): Rename to ...
(listof-ssh-host-or-ssh-match?): ... this and make more strict.
(validate-match-block): Rename to ...
(match-block?): ... this and just make it a predicate.
(serialize-ssh-host): Adjust to changes in ‘ssh-host’.
(serialize-ssh-match): Adjust to changes in ‘ssh-match’ and
‘match-block?’.
(serialize-ssh-host-or-ssh-match): Rename to ...
(serialize-listof-ssh-host-or-ssh-match): ... this.
(home-ssh-configuration): Improve docs; rename ‘host’ field to
’default-host’; and don’t use ‘serialize-toplevel-alist’.