features: emacs: matrix: add feature-matrix & minimal config
WIP, of course, and excluding pantalaimon

Q1: do you see pantalaimon having a place here, or would it be more
appropriate in a separate feature =feature-matrix=, where we can
define a home-service? this is my instinct here, to have ancillary
info for matrix defined as properties of a main feature, and we
=require= it here.

Q2: do you see a use for configuring user-mappings the interface of
this, or that of such a =feature-matrix=, similar to mail, ERC, etc?
- I have made my login as follows -- we can curry similar arguments,
  or make some config interface like those other modules, where a struct
  exists with fields for :user-id etc, and we map it into some elisp
  config object

        (let ((user "@sculp-ns:matrix.org"))
          (ement-connect
            :uri-prefix "http://localhost:8008"
            :user-id user
            :password (password-store-get (format "matrix/%s" user))))

- I don't like the define-key cons naming -- it silently (meaning
no *message* visibility) prints the alias name to the minibuffer,
which is blocking for which-key, and just a bit goofy in general.

see:

        (define-key rde-app-map (kbd "m") '("ement" . ement-list-rooms))

then load the map and watch the minibuffer

        {C-c a}

rde-applications: m = ement

...so to avoid the menu-named binding, I think it would be better to
use a prefix map here, with precedent in the app-map for `telega' --
there are many entry points to ement which can fill this prefix map,
most notably:

- "view rooms" :: (ement-list-rooms)
- "choose chat" :: (call-interactively 'ement-view-room)
- "view mentions" :: (call-interactively 'ement-notify-switch-to-mentions-buffer)
- "view notifications" :: (ement-notify-switch-to-notifications-buffer)

note that the current telega prefix map is anonymous, so we might
incur a similar silent print situation is the intention is to make
named-prefix-maps.

cheers,
  Samuel