rde: emacs: olivetti: Clarify excluding of buffers/modes
the `which-key' indirect buffer starts as 'fundamental-mode, thus
global-olivetti-mode will still apply, giving a squashed unreadable
key-suggestion buffer.

- note: we can instead check `which-key-buffer-name', or just use the
  hardcoded string

- note: updates to `global-olivetti-ignore-[modes|buffer]' variables
  may require fresh enabling of `global-olivetti-mode', if they are
  inlined during compilation

related might be the use of 'olivetti-mode' in the minibuffer, which
is useful in my 5120x1440 monitor, lest I look 45 centimeters from the
center of my screen each time I pop the minibuffer.

It's quite cool as the bindings to expand the buffer `C-c }' still
work here.

I have used this as:

    (add-hook 'minibuffer-mode-hook 'olivetti-mode)

or some variant for better width:

    (add-hook 'minibuffer-mode-hook
              (lambda ()
                (let ((olivetti-body-width 200))
                  (olivetti-mode))))