feature: emacs: Enable native-comp by default
The workaround if a package fails to build is to use the
elisp-packages-rewrites options for the rde-elisp-configuration which
carries the faulty package.

packages, Nicolas hasn't experienced any major issue. This isn't tested on all
emacs features or packages, but a workaround is provided.

Note: Besides the emacs-buttercup (patch sent) and the emacs-macrostep

Signed-off-by: Andrew Tropin <andrew@trop.in>