rde: goimapnotify: Fix pass-cmd retrieval method
Currently the command is simply the record's field. This can be #f in the case
that no pass command is specified. However in such case, features are to use
the `mail-account-get-pass-cmd` which returns the default command in that
case. This was probably accidentaly missed. Without this, goimapnotify fails
to parse it's json config, since `false` is not a valid string.

Signed-off-by: Andrew Tropin <andrew@trop.in>