features: libreoffice: Correct double initialization
This snippet now creates an error on the current guix master. The
reason is probably that the initialization option is given twice (once
through bootstraprc, once through soffice).

Removing it once seems to work.  Choosing the bootstraprc one, as it
is supposed to be more general (I'm thinking about C++ or Python
library use).

Signed-off-by: Andrew Tropin <andrew@trop.in>